// frontend/src/generated/models/Company.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Company = {
    id: string;
    name: string;
    website?: string | null;
    industry?: string | null;
    location?: string | null;
    notes?: string | null;
    archived: boolean;
    archivedAt?: string | null;
    createdAt: string;
    updatedAt: string;
    version: number;
};