// frontend/src/generated/models/CreateCompanyRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreateCompanyRequest = {
    name: string;
    website?: string | null;
    industry?: string | null;
    location?: string | null;
    notes?: string | null;
};