// frontend/src/generated/models/ProblemDetail.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProblemDetail = {
    type: string;
    title: string;
    status: number;
    detail: string;
    instance: string;
    code: string;
    timestamp: string;
    traceId: string;
    fieldErrors?: Record<string, string>;
};